requirements.txt
asgiref==3.10.0
colorama==0.4.6
Django==5.2.7
loguru==0.7.3
pillow==12.0.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
win32_setctime==1.2.0
whitenoise==6.8.2
gunicorn==23.0.0